Two way texting shell, sendMessage function and restructure the apis & apidoc

+++ src/Controller/Api/User/UserFavouriteExpertsController.php
@@ -40,8 +40,13 @@ class UserFavouriteExpertsController extends ApiController
                                                             ->contain(['Users', 'Experts'])
                                                             ->all();
 
-        $this->set('response',$userFavouriteExperts);
-        $this->set('_serialize', ['response']);
+
+        $success = true;
+        
+        $this->set('data',$userFavouriteExperts);
+        $this->set('status',$success);
+        $this->set('_serialize', ['status','data']);
+
     }
 
     /**
@@ -62,9 +67,10 @@ class UserFavouriteExpertsController extends ApiController
         ]);
 
         $success = true;
-
-        $this->set(compact('userFavouriteExpert','success'));
-        $this->set('_serialize', ['userFavouriteExpert','success']);
+        
+        $this->set('data',$userFavouriteExpert);
+        $this->set('status',$success);
+        $this->set('_serialize', ['status','data']);
     }
 
     /**
@@ -96,8 +102,9 @@ class UserFavouriteExpertsController extends ApiController
         
         $success = true;
         
-        $this->set(compact('userFavouriteExpert','success'));
-        $this->set('_serialize', ['userFavouriteExpert','success']);
+        $this->set('data',$userFavouriteExpert);
+        $this->set('status',$success);
+        $this->set('_serialize', ['status','data']);
     }
 
     /**
@@ -129,8 +136,9 @@ class UserFavouriteExpertsController extends ApiController
         
         $success = true;
         
-        $this->set(compact('userFavouriteExpert','success'));
-        $this->set('_serialize', ['userFavouriteExpert','success']);
+        $this->set('data',$userFavouriteExpert);
+        $this->set('status',$success);
+        $this->set('_serialize', ['status','data']);
     }
 
     /**
@@ -149,13 +157,17 @@ class UserFavouriteExpertsController extends ApiController
 
         $userFavouriteExpert = $this->UserFavouriteExperts->findByUserId($userId)->where(['expert_id' => $expertId])->first();
         
+        if(!$userFavouriteExpert){
+            throw new NotFoundException(__('No entity found in User favourite Experts Table'));
+        }
+
         if (!$this->UserFavouriteExperts->delete($userFavouriteExpert)) {
             throw new Exception("Expert specialization service could not be deleted.");
         }
 
         $success = true;
         
-        $this->set(compact('success'));
-        $this->set('_serialize', ['success']);
+        $this->set('status',$success);
+        $this->set('_serialize', ['status']);
     }
 }
